ALX SE

+++ 0-bubble_sort.c
@@ -8,7 +8,7 @@
 void bubble_sort(int *array, size_t size)
 {
     int *temp;
-    int i, k;
+    size_t i, k;
 
     for (i = 0; i < size; i++)
     {
